added support for adding the suffix to image urls
instead or returning a url with a hash alone, this
commit adds the ability to store the extension of
the file and also return a url with the filename as
being composed of the hash and the file extension.

This comes in handy in cases where you're using
local file storage and the web server is unable to
determine the mime-type of the file automatically.